requirements.txt
opencv-python
numpy
torch
facenet-pytorch
scikit-learn
Pillow
ultralytics
